+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/oauth/notion/api/port/NotionOauthService.java
@@ -1,7 +1,7 @@
-package io.f12.notionlinkedblog.oauth.api.port;
+package io.f12.notionlinkedblog.oauth.notion.api.port;
 
 import io.f12.notionlinkedblog.common.exceptions.exception.TokenAvailabilityFailureException;
-import io.f12.notionlinkedblog.oauth.api.response.NotionOAuthLinkDto;
+import io.f12.notionlinkedblog.oauth.notion.api.response.NotionOAuthLinkDto;
 
 public interface NotionOauthService {
