+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/oauth/notion/domain/NotionOauth.java
@@ -1,6 +1,6 @@
-package io.f12.notionlinkedblog.oauth.domain.notion;
+package io.f12.notionlinkedblog.oauth.notion.domain;
 
-import io.f12.notionlinkedblog.oauth.infrastructure.NotionOauthEntity;
+import io.f12.notionlinkedblog.oauth.notion.infrastructure.NotionOauthEntity;
 import io.f12.notionlinkedblog.user.domain.User;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
