+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/oauth/notion/domain/acceesstokendto/request/RequestAccessTokenDto.java
@@ -1,4 +1,4 @@
-package io.f12.notionlinkedblog.oauth.domain.notion.acceesstokendto.request;
+package io.f12.notionlinkedblog.oauth.notion.domain.acceesstokendto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
