+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/oauth/notion/infrastructure/NotionOauthDataRepository.java
@@ -1,10 +1,10 @@
-package io.f12.notionlinkedblog.oauth.infrastructure;
+package io.f12.notionlinkedblog.oauth.notion.infrastructure;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import io.f12.notionlinkedblog.oauth.service.port.NotionOauthRepository;
+import io.f12.notionlinkedblog.oauth.notion.service.port.NotionOauthRepository;
 
 public interface NotionOauthDataRepository extends JpaRepository<NotionOauthEntity, Long>, NotionOauthRepository {
 	Optional<NotionOauthEntity> findNotionOauthByUserId(Long userId);
