+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/oauth/notion/infrastructure/NotionOauthEntity.java
@@ -1,4 +1,4 @@
-package io.f12.notionlinkedblog.oauth.infrastructure;
+package io.f12.notionlinkedblog.oauth.notion.infrastructure;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
