+++ 003 Code/BE/notion-linked-blog/src/main/java/io/f12/notionlinkedblog/oauth/notion/service/port/NotionOauthRepository.java
@@ -1,8 +1,8 @@
-package io.f12.notionlinkedblog.oauth.service.port;
+package io.f12.notionlinkedblog.oauth.notion.service.port;
 
 import java.util.Optional;
 
-import io.f12.notionlinkedblog.oauth.infrastructure.NotionOauthEntity;
+import io.f12.notionlinkedblog.oauth.notion.infrastructure.NotionOauthEntity;
 
 public interface NotionOauthRepository {
 	Optional<NotionOauthEntity> findNotionOauthByUserId(Long userId);
